Rename project package to Texty
