+++ app/google-services.json
@@ -9,7 +9,7 @@
       "client_info": {
         "mobilesdk_app_id": "1:296282642032:android:3ff710140cf46821f586b5",
         "android_client_info": {
-          "package_name": "com.example.projectandroid"
+          "package_name": "com.example.texty"
         }
       },
       "oauth_client": [],
