--- app/src/main/java/com/example/projectandroid/model/User.kt
@@ -1,13 +1,0 @@
-package com.example.projectandroid.model
-
-/**
- * Represents a user profile stored in Firestore.
- */
-data class User(
-    val uid: String = "",
-    val displayName: String = "",
-    val photoUrl: String? = null,
-    val isOnline: Boolean = false,
-    val about: String = "",
-    val phone: String = "",
-)
