--- app/src/main/java/com/example/projectandroid/repository/UserRepository.kt
@@ -1,39 +1,0 @@
-package com.example.projectandroid.repository
-
-import com.example.projectandroid.model.User
-import com.google.firebase.firestore.FirebaseFirestore
-import com.google.firebase.firestore.ktx.firestore
-import com.google.firebase.ktx.Firebase
-
-/**
- * Repository that encapsulates user related Firestore queries.
- *
- * @param firestore allows dependency injection for testing.
- */
-class UserRepository(private val firestore: FirebaseFirestore = Firebase.firestore) {
-    private val usersCollection = firestore.collection("users")
-
-    fun getUsersByDisplayName(
-        displayName: String,
-        onSuccess: (List<User>) -> Unit,
-        onFailure: (Exception) -> Unit
-    ) {
-        if (displayName.isBlank()) {
-            onSuccess(emptyList())
-            return
-        }
-
-        usersCollection
-            .orderBy("displayName")
-            .startAt(displayName)
-            .endAt(displayName + '\uf8ff')
-            .get()
-            .addOnSuccessListener { result ->
-                val users = result.documents.mapNotNull { it.toObject(User::class.java) }
-                onSuccess(users)
-            }
-            .addOnFailureListener { exception ->
-                onFailure(exception)
-            }
-    }
-}
