--- app/src/main/java/com/example/projectandroid/ui/ChatListAdapter.kt
@@ -1,58 +1,0 @@
-package com.example.projectandroid.ui
-
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import android.widget.TextView
-import androidx.recyclerview.widget.DiffUtil
-import androidx.recyclerview.widget.ListAdapter
-import androidx.recyclerview.widget.RecyclerView
-import com.example.projectandroid.R
-import com.example.projectandroid.model.ChatRoom
-import com.google.firebase.firestore.ktx.firestore
-import com.google.firebase.ktx.Firebase
-
-class ChatListAdapter(
-    private val onClick: (ChatRoom) -> Unit,
-) : ListAdapter<ChatRoom, ChatListAdapter.ChatRoomViewHolder>(DIFF_CALLBACK) {
-
-    class ChatRoomViewHolder(view: View) : RecyclerView.ViewHolder(view) {
-        val nameText: TextView = view.findViewById(R.id.textName)
-        val lastMessageText: TextView = view.findViewById(R.id.textLastMessage)
-        val statusView: View = view.findViewById(R.id.viewStatus)
-    }
-
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): ChatRoomViewHolder {
-        val view = LayoutInflater.from(parent.context)
-            .inflate(R.layout.item_user, parent, false)
-        return ChatRoomViewHolder(view)
-    }
-
-    override fun onBindViewHolder(holder: ChatRoomViewHolder, position: Int) {
-        val room = getItem(position)
-        holder.nameText.text = room.contactName
-        holder.lastMessageText.apply {
-            text = room.lastMessage
-            visibility = View.VISIBLE
-        }
-        holder.statusView.setBackgroundResource(R.drawable.offline_indicator)
-        Firebase.firestore.collection("users").document(room.contactUid).get()
-            .addOnSuccessListener { snapshot ->
-                val online = snapshot.getBoolean("isOnline") == true
-                holder.statusView.setBackgroundResource(if (online) R.drawable.online_indicator else R.drawable.offline_indicator)
-            }
-        holder.itemView.setOnClickListener { onClick(room) }
-    }
-
-    companion object {
-        private val DIFF_CALLBACK = object : DiffUtil.ItemCallback<ChatRoom>() {
-            override fun areItemsTheSame(oldItem: ChatRoom, newItem: ChatRoom): Boolean {
-                return oldItem.id == newItem.id
-            }
-
-            override fun areContentsTheSame(oldItem: ChatRoom, newItem: ChatRoom): Boolean {
-                return oldItem == newItem
-            }
-        }
-    }
-}
