--- app/src/main/java/com/example/projectandroid/ui/UserAdapter.kt
@@ -1,48 +1,0 @@
-package com.example.projectandroid.ui
-
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import android.widget.TextView
-import androidx.recyclerview.widget.DiffUtil
-import androidx.recyclerview.widget.ListAdapter
-import androidx.recyclerview.widget.RecyclerView
-import com.example.projectandroid.R
-import com.example.projectandroid.model.User
-
-class UserAdapter(
-    private val onClick: (User) -> Unit
-) : ListAdapter<User, UserAdapter.UserViewHolder>(DIFF_CALLBACK) {
-
-    class UserViewHolder(view: View) : RecyclerView.ViewHolder(view) {
-        val nameText: TextView = view.findViewById(R.id.textName)
-        val statusView: View = view.findViewById(R.id.viewStatus)
-    }
-
-    override fun onCreateViewHolder(parent: ViewGroup, viewType: Int): UserViewHolder {
-        val view = LayoutInflater.from(parent.context)
-            .inflate(R.layout.item_user, parent, false)
-        return UserViewHolder(view)
-    }
-
-    override fun onBindViewHolder(holder: UserViewHolder, position: Int) {
-        val user = getItem(position)
-        holder.nameText.text = user.displayName
-        holder.statusView.setBackgroundResource(
-            if (user.isOnline) R.drawable.online_indicator else R.drawable.offline_indicator
-        )
-        holder.itemView.setOnClickListener { onClick(user) }
-    }
-
-    companion object {
-        private val DIFF_CALLBACK = object : DiffUtil.ItemCallback<User>() {
-            override fun areItemsTheSame(oldItem: User, newItem: User): Boolean {
-                return oldItem.uid == newItem.uid
-            }
-
-            override fun areContentsTheSame(oldItem: User, newItem: User): Boolean {
-                return oldItem == newItem
-            }
-        }
-    }
-}
