+++ app/src/main/java/com/example/texty/model/User.kt
@@ -1,0 +1,13 @@
+package com.example.texty.model
+
+/**
+ * Represents a user profile stored in Firestore.
+ */
+data class User(
+    val uid: String = "",
+    val displayName: String = "",
+    val photoUrl: String? = null,
+    val isOnline: Boolean = false,
+    val about: String = "",
+    val phone: String = "",
+)
