+++ app/src/main/res/values/strings.xml
@@ -1,5 +1,5 @@
 <resources>
-    <string name="app_name">projectAndroid</string>
+    <string name="app_name">Texty</string>
     <string name="invalid_email">Email inválido</string>
     <string name="invalid_password">La contraseña debe tener al menos 6 caracteres</string>
     <string name="error_generic">Error</string>
